Moved configSync validation to chromeStorage.js
